[H2O] Moved the client java file to new location
