+++ benchmarks/bms/h2o/src/org/dacapo/h2o/DacapoH2OFacade.java
@@ -1,4 +1,4 @@
-package water;
+package org.dacapo.h2o;
 
 import sun.security.provider.MD5;
