+++ benchmarks/bms/h2o/src/org/dacapo/h2o/RestUtil.java
@@ -1,4 +1,4 @@
-package water;
+package org.dacapo.h2o;
 
 import net.sf.json.JSONObject;
